Agregando rolUsuario en tabla Usuario

+++ ProyectoEJB-ejb/src/java/com/fannog/proyectoservidor/entities/Analista.java
@@ -27,7 +27,7 @@ public class Analista extends Usuario implements Serializable {
     @OneToMany(mappedBy = "analista")
     private List<Constancia> constancias;
 
-    public Analista(String apellidos, String documento, String email, String nombres, Integer telefono, String password, EstadoUsuario estado, Localidad localidad, Integer rolUsuario) {
+    public Analista(String apellidos, String documento, String email, String nombres, Integer telefono, String password, EstadoUsuario estado, Localidad localidad, Long rolUsuario) {
         super(apellidos, documento, email, nombres, telefono, password, estado, localidad, rolUsuario);
     }
